+++ backend/src/main/java/com/example/health_care/entity/RecoveryQuestionCode.java
@@ -1,0 +1,8 @@
+package com.example.health_care.entity;
+
+public enum RecoveryQuestionCode {
+    BIRTHPLACE,         // 내가 태어난 곳은?
+    ELEMENTARY_SCHOOL,  // 내가 다니던 초등학교는?
+    PET_NAME,           // 내 애완동물 이름은?
+    MOTHER_NAME         // 내 어머니 이름은?
+}
