+++ backend/src/main/resources/db/recovery.sql
@@ -3,13 +3,23 @@
 -- 파일명: Recovery.sql
 -- 테이블명: recovery
 -- ==========================
+-- 기존 테이블이 있다면 삭제
+DROP TABLE recovery CASCADE CONSTRAINTS;
+
 CREATE TABLE recovery (
     idx           NUMBER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    customer_id NUMBER NOT NULL,  -- customers.id (이메일)
-    code          VARCHAR2(32)    NOT NULL,  -- enum: BIRTHPLACE / ELEMENTARY_SCHOOL / PET_NAME / MOTHER_NAME
+    -- customers 테이블의 기본 키인 idx(NUMBER)와 동일한 데이터 타입으로 변경
+    customer_id   NUMBER          NOT NULL,
+    -- 고객 식별자가 idx(NUMBER)를 참조하도록 설정
+    code          VARCHAR2(32)    NOT NULL,
     answer_hash   VARCHAR2(255)   NOT NULL,
     updated_at    TIMESTAMP       NOT NULL,
-    CONSTRAINT uq_recovery_user_code UNIQUE (customer_id, code)
+    
+    -- 고객 식별자(idx)와 질문 코드를 사용하여 고유 제약 조건 설정
+    CONSTRAINT uq_recovery_user_code UNIQUE (customer_id, code),
+    
+    -- customers 테이블의 기본 키(idx)를 참조하는 외래 키 설정
+    FOREIGN KEY (customer_id) REFERENCES customers(idx) ON DELETE CASCADE
 );
 
 -- 필요 시 FK를 customers(id)에 연결 (고객 테이블 설계가 id UNIQUE 인 환경에서만 권장)
